requirements.txt
Pillow
pymupdf
numpy